requirements.txt
azure-functions
python-dotenv
openai>=1.0.0
google-genai>=0.3.0
fastapi
uvicorn
python-dotenv